+++ packages/app/src/SmartSearchPage.test.tsx
@@ -1,11 +1,10 @@
 import { PropertyType } from '@medplum/core';
 import { MockClient } from '@medplum/mock';
-import { FhirPathTableField, MedplumProvider } from '@medplum/react';
+import { FhirPathTableField, Loading, MedplumProvider } from '@medplum/react';
 import { act, fireEvent, render, screen, waitFor } from '@testing-library/react';
 import React, { Suspense } from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { AppRoutes } from './AppRoutes';
-import { Loading } from './components/Loading';
 
 const query = `{
 ResourceList: ServiceRequestList {
