+++ packages/app/src/resource/ResourceVersionPage.tsx
@@ -1,9 +1,8 @@
 import { Paper, Tabs, Text, Title } from '@mantine/core';
 import { Bundle, BundleEntry, OperationOutcome, Resource, ResourceType } from '@medplum/fhirtypes';
-import { Container, Document, MedplumLink, ResourceDiff, useMedplum } from '@medplum/react';
+import { Container, Document, Loading, MedplumLink, ResourceDiff, useMedplum } from '@medplum/react';
 import React, { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Loading } from '../components/Loading';
 
 export function ResourceVersionPage(): JSX.Element {
   const navigate = useNavigate();
