+++ packages/react/src/index.ts
@@ -1,21 +1,22 @@
 export * from './AddressDisplay/AddressDisplay';
 export * from './AddressInput/AddressInput';
 export * from './AnnotationInput/AnnotationInput';
+export * from './AppShell/AppShell';
+export * from './AppShell/Header';
+export * from './AppShell/Navbar';
 export * from './AsyncAutocomplete/AsyncAutocomplete';
 export * from './AttachmentArrayDisplay/AttachmentArrayDisplay';
 export * from './AttachmentArrayInput/AttachmentArrayInput';
 export * from './AttachmentButton/AttachmentButton';
 export * from './AttachmentDisplay/AttachmentDisplay';
 export * from './AttachmentInput/AttachmentInput';
-export * from './auth/RegisterForm';
-export * from './auth/SignInForm';
 export * from './BackboneElementDisplay/BackboneElementDisplay';
 export * from './BackboneElementInput/BackboneElementInput';
 export * from './CalendarInput/CalendarInput';
 export * from './CheckboxFormSection/CheckboxFormSection';
+export * from './CodeInput/CodeInput';
 export * from './CodeableConceptDisplay/CodeableConceptDisplay';
 export * from './CodeableConceptInput/CodeableConceptInput';
-export * from './CodeInput/CodeInput';
 export * from './CodingDisplay/CodingDisplay';
 export * from './CodingInput/CodingInput';
 export * from './ContactDetailDisplay/ContactDetailDisplay';
@@ -38,6 +39,7 @@ export * from './HumanNameDisplay/HumanNameDisplay';
 export * from './HumanNameInput/HumanNameInput';
 export * from './IdentifierDisplay/IdentifierDisplay';
 export * from './IdentifierInput/IdentifierInput';
+export * from './Loading/Loading';
 export * from './Logo/Logo';
 export * from './MedplumLink/MedplumLink';
 export * from './MedplumProvider/MedplumProvider';
@@ -80,6 +82,8 @@ export * from './ServiceRequestTimeline/ServiceRequestTimeline';
 export * from './StatusBadge/StatusBadge';
 export * from './Timeline/Timeline';
 export * from './TimingInput/TimingInput';
+export * from './auth/RegisterForm';
+export * from './auth/SignInForm';
 export * from './useResource/useResource';
 export * from './utils/date';
 export * from './utils/outcomes';
